Camera recorder (#38)
* feat: added pool of buffers and timers

* feat: added image queue handler and syncronization logic

* feat: signal handling and main function added
fix: improved comments in queue synchronization

* feat: completed synchronization without saving, texted below 85 FPS

* feat: added lz4 cmake configuration and ready for mcap integration

* feat: done with mmaping files

* fix: fixed file size to be able to read it, clear debug output

* fix: seg faults on writing to mmaped file
no-fix: lack of buffers on 100 FPS

* fix: reworked remapping multithread synchronization
fix: added atomics
todo: remove debug and output lines

* fix: removed debug output and unneccessary variables

+++ packages/camera/src/calibration.cpp
@@ -163,7 +163,7 @@ void CalibrationNode::clearLastDetection(int camera_idx) {
 void CalibrationNode::fillImageObjectPoints(
     std::vector<std::vector<cv::Point2f>>& image_points,
     std::vector<std::vector<cv::Point3f>>& obj_points, int camera_idx) {
-    for (int i = 0; i < state_.marker_corners_all[camera_idx].size(); ++i) {
+    for (size_t i = 0; i < state_.marker_corners_all[camera_idx].size(); ++i) {
         image_points.emplace_back();
         obj_points.emplace_back();
         param_.aruco_board.matchImagePoints(
